condaenv.zw24042s.requirements.txt
addict==2.4.0
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
beautifulsoup4==4.13.3
bresenham==0.2.1
cachetools==5.5.2
colorama==0.4.6
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
cython==3.0.12
distro==1.9.0
et-xmlfile==2.0.0
eval-type-backport==0.2.2
fire==0.7.0
fonttools==4.57.0
imageio==2.35.1
importlib-resources==6.4.5
jiter==0.9.0
kiwisolver==1.4.7
lazy-loader==0.4
lmdb==1.6.2
lxml==5.3.2
matplotlib==3.7.5
mmcv-full==1.7.1
mmdet==2.28.2
more-itertools==10.5.0
numpy==1.24.4
openai==1.74.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pandas==2.0.3
pdf2image==1.17.0
platformdirs==4.3.6
premailer==3.10.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.7
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pywavelets==1.4.1
pyyaml==6.0.2
rapidfuzz==3.9.7
regex==2024.11.6
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.7
soupsieve==2.6
termcolor==2.4.0
terminaltables==3.1.10
tifffile==2023.7.10
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.99
ultralytics-thop==2.0.14
yapf==0.43.0
zipp==3.20.2